ref: holder of collection requirements
