+++ components/collection/drop/Generative.vue
@@ -1,8 +1,6 @@
 <template>
   <CollectionDropGenerativeLayout
     :user-minted-nft-id="userMintedNftId"
-    :user-minted-count="mintedAmountForCurrentUser"
-    :user-max-available-to-mint="1"
     :is-wallet-connecting="isWalletConnecting"
     :is-image-fetching="isImageFetching"
     :is-loading="isLoading"
@@ -144,7 +142,6 @@ const {
   canListMintedNft,
   mintedCount,
   mintCountAvailable,
-  mintedAmountForCurrentUser,
   selectedImage,
   description,
   collectionName,
