+++ components/collection/drop/PaidGenerative.vue
@@ -4,8 +4,6 @@
     :description="description"
     :drop="drop"
     :user-minted-nft-id="userMintedNftId"
-    :user-max-available-to-mint="Infinity"
-    :user-minted-count="mintedAmountForCurrentUser"
     :is-wallet-connecting="isWalletConnecting"
     :is-image-fetching="isImageFetching"
     :is-loading="isLoading"
